pyproject.toml
[project]
name = "mygit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.44",
    "pyqt6>=6.8.1",
]
